Modified validations

+++ reservation/src/main/java/com/accelerate/acceleronServices/reservation/service/ReservationRequestServiceImpl.java
@@ -14,6 +14,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Optional;
 
@@ -32,43 +33,53 @@ public class ReservationRequestServiceImpl implements ReservationRequestService 
     @Transactional
     public ApiResponse<GenericResponse> makeReservation(ReservationRequestDto request) {
 
-        reservationRepository.save(entityDtoConversion.convertToEntity(request));
+        ReservationEntity reservationEntity = entityDtoConversion.convertToEntity(request);
 
-        return new ApiResponse<>().successResponse("Success");
+        reservationRepository.save(reservationEntity);
+
+        ApiResponse<GenericResponse> response = new ApiResponse<>(true,HttpStatus.OK.value(),"Reservation Created",
+                new GenericResponse(StatusTextEnum.SUCCESS.value()));
+
+        return response;
 
     }
     @Override
     public List<ReservationEntity> getAllReservation(String search, Integer limit, Integer skip){
 
-        if(search==null) {  //search is not provided
-            if (limit == null && skip == null) {
-                return reservationRepository.findAll();  //return all table rows
-            }
-            if (limit != null && skip == null) {
-                return reservationRepository.findAllById(limit, 0);  //returns first x rows where x=limit
-            }
-            return reservationRepository.findAllById(limit, skip);
+        if (limit == null && skip == null) {
+            return reservationRepository.findAll();  //return all table rows
         }
-        else{  //search is provided
-            if(skip==null){
-                skip=0;
-            }
-            if(limit==null){
-                limit=10;
-            }
-            return reservationRepository.findByGivenString(search,limit,0);
+        if (limit != null && skip == null) {
+            return reservationRepository.findAllById(limit, 0);  //returns first x rows where x=limit
         }
+        return reservationRepository.findAllById(limit, skip);
     }
     @Override
-    public Optional<ReservationEntity> getReservationById(int id){
-        return reservationRepository.findById(id);
+    public ReservationEntity getReservationById(int id) {
+        ReservationEntity reservationEntity = reservationRepository.findById(id);
+        if (reservationEntity != null) {
+            return reservationEntity;
+        } else {
+            throw new EntityNotFoundException("Entity not found for the id " + id);
+        }
     }
+
 
     @Override
     public ApiResponse<GenericResponse> deleteReservation(int id) {
 
-        reservationRepository.deleteById(id);
-        return new ApiResponse<>().successResponse("Deleted");
+        if (reservationRepository.existsById(id)){
+            reservationRepository.deleteById(id);
+            ApiResponse<GenericResponse> response = new ApiResponse<>(true,HttpStatus.OK.value(),"Reservation Deleted",
+                    new GenericResponse(StatusTextEnum.SUCCESS.value()));
+            return response;
+        }
+        else{
+            //if the id is not present
+            throw new EntityNotFoundException("Entity not found for the id " + id);
+        }
+
+
     }
 
 
@@ -76,15 +87,24 @@ public class ReservationRequestServiceImpl implements ReservationRequestService 
     public ApiResponse<GenericResponse> updateReservation(int id, ReservationRequestDto request) {
 
         ReservationEntity requestEntity = entityDtoConversion.convertToEntity(request);
-        Optional<ReservationEntity> reservationEntity = reservationRepository.findById(id);
+        ReservationEntity reservationEntity = reservationRepository.findById(id);
 
-        if(reservationEntity.isPresent()){
-            requestEntity.setId(reservationEntity.get().getId());
+        if(reservationEntity != null){
+            requestEntity.setId(reservationEntity.getId());
             reservationRepository.save(requestEntity);
-            return new ApiResponse<>().successResponse("Updated your reservation");
+            return new ApiResponse<>(true,HttpStatus.OK.value(),"Reservation Created",
+                    new GenericResponse(StatusTextEnum.SUCCESS.value()));
         }
         else{ //if the id is not present
-            return new ApiResponse<>().successResponse("The given id doesn't exist");
+            throw new EntityNotFoundException("Entity not found for the id " + id);
         }
+    }
+
+    @Override
+    public List<ReservationEntity> getAllReservationBySearch(String search, Integer limit) {
+        if(limit==null){
+            limit=10;
+        }
+        return reservationRepository.findByGivenString(search,limit,0);
     }
 }
